+++ Assets/Packages/com.dman.grid-game-tools/Tests/WorldBuilding/WorldBuilder.cs
@@ -40,7 +40,7 @@ namespace GridDomain.Test
         /// build to a world with certain values set to defaults
         /// </summary>
         /// <returns></returns>
-        public IDungeonWorld BuildToWorld(WorldBuildString characterMap, ulong seed = 0, IEnumerable<IWorldComponent> components = null)
+        public IDungeonWorld BuildToWorld(WorldBuildString characterMap, uint seed = 0, IEnumerable<IWorldComponent> components = null)
         {
             var allEntities = Build(Vector3Int.zero, characterMap);
             var bounds = new DungeonBounds(Vector3Int.zero, characterMap.Size());
